SDK cache: deconstruct the default cache value, don't use that object
Problem:
const ethSdk = new CegaEvmSDK(...)
const arbSdk = new CegaEvmSDK(...)

it was using the same object as the cache, and so if we updated the
cache on ethereum, it would also update on arb side. This would mess up
some of the numbers.

Solution:
Deconstruct the default cache value creates a new object in the place of
the cache so there should be no relationship between each instance of
the sdk.